+++ elements/user-picker/src/__tests__/visual-regression/_constants.ts
@@ -1,0 +1,16 @@
+export const CLEAR_INDICATOR_SELECTOR = '.fabric-user-picker__clear-indicator';
+
+export const CONTROL_SELECTOR = '.fabric-user-picker__control';
+
+export const DISABLE_INPUT_ID_MULTI = '#disabled-input-multi-user-picker';
+
+export const DISABLE_INPUT_ID_SINGLE = '#disabled-input-single-user-picker';
+
+export const INPUT_SELECTOR = '#react-select-example-input';
+
+export const LOZENGE_CLEAR_ICON_SELECTOR =
+  '.fabric-user-picker__multi-value__remove';
+
+export const LOZENGE_CLEAR_INDICATOR_SELECTOR = 'show-clear-icon';
+
+export const POPUP_BUTTON = '[data-testid="popup-button-0"]';