+++ in-product-testing/in-product-testing-sample/__tests__/in-product/tests/notifications.ts
@@ -1,0 +1,13 @@
+import {
+  notificationDocumentToggleTestCase,
+  notificationRendersTestCase,
+} from '@atlassian/notification-common/in-product';
+
+describe('Notifications', () => {
+  beforeEach(() => {
+    cy.navigateTo('notifications', 'notification-common', 'basic');
+  });
+
+  notificationRendersTestCase().test(cy);
+  notificationDocumentToggleTestCase().test(cy);
+});